+++ package.json
@@ -82,8 +82,8 @@
     "build-storybook": "build-storybook",
     "dev": "NODE_ENV=development npm start",
     "lint": "./node_modules/.bin/eslint src",
-    "sim": "./node_modules/.bin/babel-node src/openstack-simulator/main.js",
-    "start": "./node_modules/.bin/babel-node src/server/server.js",
+    "start": "./node_modules/.bin/babel-node src/devserver/server.js",
+    "server": "./node_modules/.bin/babel-node src/server/main.js",
     "storybook": "start-storybook -p 9001 -c src/stories",
     "test": "npm run lint && npm run test:unit",
     "test:e2e": "NODE_ENV=test jest -i e2e-tests/",
