+++ src/server/api/keystone/deleteProject.js
