+++ src/server/api/neutron/deleteNetwork.js
