+++ src/server/api/nova/deleteFlavor.js
