+++ src/server/api/nova/getFlavors.js
